requirements.txt
Flask==3.0.3
pytorch==2.1.2
tqdm==4.66.2